Implement Vector (r=edwsmith)
This patch implements the Vector class from the ES4 proposal.  Note that it
does not implement type-parameters, expecting the compiler to turn the
type-parameter syntax for Vectors into concrete classes.  These concrete
classes are optimized for int, uint, and Number.  Once type-parameters are
implemented, we can go back and redo Vector that way.

The JIT has been modified to call fast get/set methods on the Vector classes.
This is similar to the optimizations done for Array.  The difference here is
that for int, uint, or Number Vectors we know the specific type of the
operation and can avoid unnecessary conversion to Atom and back (and between
int and double when dealing with large ints).

This adds 4 new builtin types (Vector$int, Vector$uint, Vector$double, and
Vector$object for the general case).

I also added the errorgen script and neccessary .xml files that were missing
from the hg repo.  I think these got left out from the initial cvs checkin, and
no one has had to add error messages since.