Support for profiling JIT'ed Javascript/ActionScript code under VTune.
A windows project file, VTune, was added in order to build
a version of the shell that has VTune support built in.

You must also have the 'JIT vtune patch' available in modules/vtune/
(at the same level as the zlib directory.  See
http://developer.mozilla.org/en/docs/Tamarin_Build_Documentation)
This is where JITProfiling.h and the libraries are pulled from.

The support consists of a call from AvmCore which determines
whether the run is under VTune, exposing some fields under
a 'release' configuration that were previously only available
in 'debugger' builds, and registering methods to vtune during
code generation.

Code generation was also modified to support calling out
to vtune during prologue and epilogue for call graphs.

Kudos to Mohammad Haghighat, Shengnan Cong and Marsha Eng
for making this patch possible.